Configure Jackson's object mapper to work with the JVM time zone
Without any configuration Jackson falls back to work with a UTC-based
time correction mechanism. While this won't affect date time fields, it
does get in the way when we are working with date-only fields.
Previously, before serialization each date would be translated to its
UTC bound equivalent which, depending on the time zone of the JVM, could
potentially shift the date to a day earlier. A similar bug can occur on
deserialization.