Set @JsonIgnore on Dokumentversjon getters
This is needed because Jackson is configured to serialize fields based
on the getters and the @JsonIgnore annotations we had on the fields
themselves were actually ignored since the dawn of time. This led to
these fields being sent to the backend when a document version was
updated which was, in turn, rejected by said backend which does not
allow these fields to be updated.